+++ src/main.rs
@@ -1,25 +1,22 @@
 #![no_std]
 #![feature(start)]
 
-mod gba_color;
-mod graphics;
+mod color;
 mod picture;
-mod rgb;
 mod screen;
+mod vram;
 
-use graphics::Graphics;
 use picture::SAMPLE_PICT;
 use screen::Screen;
-// use rgb::RGBDef;
-// use rgb::RGB;
+use vram::Vram;
 
 #[start]
 fn main(_argc: isize, _argv: *const *const u8) -> isize {
     let screen = Screen::new();
     screen.mode3_init();
 
-    let graphics: Graphics = Graphics::new();
-    graphics.paint_pict(SAMPLE_PICT);
+    let vram: Vram = Vram::new();
+    vram.paint_pict(SAMPLE_PICT);
 
     loop {}
 }
